Refactor GLTF export extensions to use overrideExporterExtension
